Move newLineFormatter to prescription component dir
