+++ tool/site/client/src/components/prescription-summary/prescriptionLevelDetails.tsx
@@ -3,7 +3,7 @@ import React, {FC} from "react"
 import {CommunicationRequest, CommunicationRequestPayload, MedicationRequest} from "fhir/r4"
 import {formatCurrentDate, formatDate} from "../../formatters/dates"
 import {getPerformerSiteTypeExtension} from "../../fhir/customExtensions"
-import {newLineFormatter} from "./newLineFormatter"
+import {newLineFormatter} from "../prescription/newLineFormatter"
 import {COURSE_OF_THERAPY_TYPE_CODES, VALUE_SET_COURSE_OF_THERAPY_TYPE} from "../../fhir/reference-data/valueSets"
 import {getCurrentIssueNumberAndEndIssueNumber} from "../../fhir/helpers"
 import {Field} from "formik"
